realização de exercicios sobre Casting

+++ src/EBACJava/Modulo02/Domain/Casting.java
@@ -2,5 +2,5 @@ package EBACJava.Modulo02.Domain;
 
 public class Casting {
     int num1 = 10;
-    short num2 = num1;
+    short num2 = (short) num1;
 }
